change LayerType and rm relations
